% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{default}
\alias{default}
\alias{resgf_get_indexNode}
\alias{resgf_set_indexNode}
\title{Get the address of the default index node}
\usage{
resgf_get_indexNode()

resgf_set_indexNode(node)
}
\arguments{
\item{node}{URL (including http://) of the default index node to be used}
}
\value{
A text url of the currently configured default index node
}
\description{
Get the address of the default index node
}
\details{
This value is stored as the option `resgf.indexNode`
}
