% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resGF.R
\name{resGF}
\alias{resGF}
\title{Generate resistance surface from GF analysis}
\usage{
resGF(obj, raster_stack, save.image = TRUE, results_dir)
}
\arguments{
\item{obj}{A GF object}

\item{raster_stack}{A raster stack oject containing the same variable used in the GF analysis}

\item{save.image}{Save individual conversion of landscape variables into resistance surfaces}

\item{results_dir}{Directory to save the transformation images}
}
\value{
A resistance surface
}
\description{
This function generate a resistance surface based on gradient forest analysis.
}
