% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{resgf_build_constraints}
\alias{resgf_build_constraints}
\title{Builds a set of search constraints for use with ESGF restful API.}
\usage{
resgf_build_constraints(...)
}
\arguments{
\item{...}{Constraints supplied as a named list}
}
\description{
Internal function
}
