% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.r
\name{resgf_check_status}
\alias{resgf_check_status}
\title{Compare local database with ESGF search}
\usage{
resgf_check_status(local.db.dir, esgf.manifest, check.checksums = FALSE)
}
\arguments{
\item{local.db.dir}{Path to the local directory. Files nested in subdirectories are also detected.}

\item{esgf.manifest}{resgf_manifest object generated previously, which forms the reference database}

\item{check.checksums}{Should the comparison be done using the checksums (slow!)? Or only the filenames.}
}
\value{
An resgf_status object
}
\description{
Compares the list of files ("manifest") generated by an ESGF search with the locally downloaded files
}
\examples{
resgf_check_status(local.db.dir=pwd(),
                   esgf.manifest=this.manifest)
}
