% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{resgf_extract_manifest}
\alias{resgf_extract_manifest}
\title{Extract the manifest from a wget script}
\usage{
resgf_extract_manifest(wget.script)
}
\arguments{
\item{wget.script}{The contents of the wget script as a single character vector}
}
\value{
A tibble with the manifest
}
\description{
Extract the manifest from a wget script
}
