% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{resgf_get_filelist}
\alias{resgf_get_filelist}
\title{Get the filelist}
\usage{
resgf_get_filelist(object, max.files = 1000, max.datasets = 10)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Gets the list of files associated with the datasets
}
