% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{resgf_get_manifest}
\alias{resgf_get_manifest}
\title{Get the manifest of files from an ESGF search}
\usage{
resgf_get_manifest(..., node = "http://esgf-node.llnl.gov/esg-search")
}
\arguments{
\item{...}{A list of constraints on which to apply the search. See the ESGF RESTful API documentation for help, https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html}

\item{node}{URL (including "http://") of the ESGF index node to search}
}
\value{
An resgf_manifest object containing the manifest
}
\description{
Get the manifest of files from an ESGF search
}
\examples{
resgf_get_manifest(variable_id=c("tos","so"),
                   experiment_id="historical",
                   table_id="Omon",
                   project="CMIP6")
}
