% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{resgf_search}
\alias{resgf_search}
\title{Search ESGF}
\usage{
resgf_search(
  ...,
  node = "http://esgf-node.llnl.gov/esg-search",
  search.limit = 10000
)
}
\arguments{
\item{...}{A list of constraints on which to apply the search. See the ESGF RESTful API documentation for help, https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html}

\item{node}{URL (including "http://") of the ESGF index node to search}

\item{search.limit}{Maximum number of values to return in the search. ESGF currently limits this to 10000}
}
\value{
A tibble detailing the search results
}
\description{
Search ESGF
}
\examples{
resgf_search(variable_id=c("tos","so"),
             experiment_id="historical",
             table_id="Omon",
             project="CMIP6")
}
