% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{resgf_simplify}
\alias{resgf_simplify}
\title{Simplify an resgf* object}
\usage{
resgf_simplify(object)
}
\arguments{
\item{object}{Raw object to simplify}
}
\value{

}
\description{
The datasets returned by searching resgf don't always parse conveniently into a tibble - some columns end up as
lists, rather than vectors. This function simplifies and makes that conversion where it can.
}
