% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.r
\name{resgf_status_check}
\alias{resgf_status_check}
\title{Compare local database with ESGF search}
\usage{
resgf_status_check(local.dir, remote.db, check.checksums = FALSE)
}
\arguments{
\item{local.dir}{Path to the local directory. Files nested in subdirectories are also detected.}

\item{remote.db}{resgf_fileset object generated previously, which forms the reference database}

\item{check.checksums}{Should the comparison be done using the checksums (slow!)? Or only the filenames.}
}
\value{
An resgf_status object
}
\description{
Compares the list of files generated by an ESGF search with the locally downloaded files
}
\examples{
resgf_check_status(local.db.dir=pwd(),
                   esgf.manifest=this.manifest)
}
