% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.r
\name{summariseStatus}
\alias{summariseStatus}
\alias{resgf_status_by_field}
\alias{resgf_status_by_node}
\title{Summarise status}
\usage{
resgf_status_by_field(object, field.idx = 1, field.sep = "_")

resgf_status_by_node(object)
}
\arguments{
\item{object}{An resgf_status object}

\item{field.idx}{Index of the filename field to summarise over.}

\item{field.sep}{Character used to separate fields in the filename. Default is "_"}
}
\value{
A tibble summarising the status accordingly.
}
\description{
Summarise the results stored in a resgf_status object in terms of one of the filename fields or the data node
}
