% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{resgf_get_filelist}
\alias{resgf_get_filelist}
\title{Get the filelist}
\usage{
resgf_get_filelist(
  object,
  max.files = 1000,
  max.datasets = 10,
  processes = 1,
  index.node = getOption("resgf.indexNode")
)
}
\arguments{
\item{object}{resgfDataset object from which to retrieve the underlying files}

\item{max.files}{maximum number of files to be returned in a chunk}

\item{max.datasets}{maximum number of datasets to request in a chunk}

\item{processes}{Number of processes to perform in parallel}

\item{index.node}{URL (including "http://") of the ESGF index node to search. Defaults to options("resgf.indexNode"). 
Passed in this instance further to `resgf_search*()`}
}
\value{

}
\description{
Gets the list of files associated with the datasets. This is done in a piecewise manner to avoid
overloading both the number of return values and the ability to issue the command, and can optionally
be done in parallel as well.
}
