% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.r
\name{resgf_retrieve}
\alias{resgf_retrieve}
\title{Retrieve missing files from ESGF}
\usage{
resgf_retrieve(
  object,
  node = "http://esgf-node.llnl.gov/esg-search",
  processes = 1
)
}
\arguments{
\item{object}{Either a resgf_status or resgf_manifest object detailing what should be retrieved}

\item{node}{ESGF node to generate wget scripts from}

\item{processes}{Number of processes to run in parallel}
}
\description{
Retrieve missing files from ESGF
}
\details{
When a resgf_status object is supplied, only files that are not held locally (but are listed in the manifest) are retrieved. For a manifest object, all files are retrieved
are
}
