% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.r
\name{resgf_retrieve}
\alias{resgf_retrieve}
\title{Retrieve missing files from ESGF}
\usage{
resgf_retrieve(
  object,
  local.dir = "missing",
  node = "http://esgf-node.llnl.gov/esg-search",
  processes = 1,
  keep.script = FALSE
)
}
\arguments{
\item{object}{Either a resgfStatus or resgfFileset object detailing what should be retrieved}

\item{local.dir}{The local directory in which to download - overrides the default taken from an resgfStatus object.}

\item{node}{ESGF node to generate wget scripts from}

\item{processes}{Number of processes to run in parallel}

\item{keep.script}{Retains the wget download script after completion.}
}
\description{
Retrieve missing files from ESGF
}
\details{
When a resgfStatus object is supplied, only files that are not held locally or that have failed checksum checks 
(if performed)are retrieved. For a resgfFileset object, all files are retrieved.
}
