% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.r
\name{resgf_retrieve}
\alias{resgf_retrieve}
\title{Retrieve missing files from ESGF}
\usage{
resgf_retrieve(object)
}
\arguments{
\item{object}{Either a resgf_status or resgf_manifest object detailing what should be retrieved}
}
\description{
Retrieve missing files from ESGF
}
\details{
When a resgf_status object is supplied, only files that are not held locally (but are listed in the manifest) are retrieved. For a manifest object, all files are retrieved
are
}
